Rename actuators -> devices
