+++ devices.py
